Further investigation into which usercontroller functions needs to be private

+++ bb-server/src/controllers/UserController.ts
@@ -25,7 +25,8 @@ export default class UserController {
 	}
 
 	@Query()
-	public userAuthenticated(): User {
+	public userAuthenticated(): User | undefined {
+		if (this.currentUser === undefined) { return undefined; }
 		return this.currentUser;	// Will return undefined if auth fails
 	}
 
@@ -42,32 +43,38 @@ export default class UserController {
 
 	@Mutation()
 	public async userEdit(args: IInput<IUserEditArgs>): Promise<User | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		return this.userRepository.save({ ...this.currentUser, ...args.input });
 	}
 
 	@Mutation()
-	public async userDelete(): Promise<boolean> { // Maybe prompt on front-end first
+	public async userDelete(): Promise<boolean | undefined> { // Maybe prompt on front-end first
+		if (this.currentUser === undefined) { return undefined; }
 		this.userRepository.remove(this.currentUser);
 		return true;
 	}
 
 	@Mutation()
 	public userToggleWhitelist(args: IUserToggleTopiclistArgs): Promise<Topic[] | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		return this.userRepository.toggleWhitelist(this.currentUser, args.topics);
 	}
 
 	@Mutation()
 	public userToggleBlacklist(args: IUserToggleTopiclistArgs): Promise<Topic[] | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		return this.userRepository.toggleBlacklist(this.currentUser, args.topics);
 	}
 
 	@Mutation()
 	public userToggleFollowedTags(args: IUserToggleTagsArgs): Promise<Tag[] | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		return this.userRepository.toggleFollowedTags(this.currentUser, args.tags);
 	}
 
 	@Mutation()
 	public async userToggleFollowing(args: IUserArgs): Promise<User[] | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		const fullUser: User = await this.userRepository.findOne(this.currentUser.id, { relations: ["followedUsers"] });
 		const toFollow: User | undefined = await this.userRepository.findOne(args.id);
 		if (toFollow === undefined) { return undefined; }
@@ -79,6 +86,7 @@ export default class UserController {
 
 	@Mutation()
 	public async userToggleSavedRecipe(args: IUserArgs): Promise<Recipe[] | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		const fullUser: User = await this.userRepository.findOne(this.currentUser.id, { relations: ["savedRecipes"] });
 		const toSave: Recipe | undefined = await this.recipeRepository.findOne(args.id);
 		if (toSave === undefined) { return undefined; }
@@ -89,7 +97,8 @@ export default class UserController {
 	}
 
 	@Mutation()
-	public async userReviewSave(args: IInput<IUserReviewSaveArgs>): Promise<UserReview> {
+	public async userReviewSave(args: IInput<IUserReviewSaveArgs>): Promise<UserReview | undefined> {
+		if (this.currentUser === undefined) { return undefined; }	// Need to check if user was in a past attended meal
 		const review: UserReview | undefined = await this.userReviewRepository.findOne({
 			subject: {
 				id: args.input.subjectID,
@@ -112,7 +121,13 @@ export default class UserController {
 
 	@Mutation()
 	public async userReviewEdit(args: IInput<IUserReviewEditArgs>): Promise<UserReview | undefined> {
+		if (this.currentUser === undefined) { return undefined; }
 		const review: UserReview | undefined = await this.userReviewRepository.findOne(args.input.id);
+
+		if (this.currentUser.id !== review.author.id) {
+			return undefined;
+		}
+
 		return review === undefined ? undefined : this.userReviewRepository.save(
 			{
 				...review,
